+ Changes to TestAero: added methods to compute number of turns at safe and max thrust.  Also adjusted the getFuelTonnage method signature for Aero, now takes a float instead of an int.

+++ trunk/megamek/src/megamek/common/Aero.java
@@ -646,7 +646,7 @@ public class Aero extends Entity {
      * 
      * @param fuelTons  The number of tons of fuel
      */
-    public void setFuelTonnage(int fuelTons){
+    public void setFuelTonnage(float fuelTons){
         float pointsPerTon = getFuelPointsPerTon();        
         fuel = (int)Math.ceil(pointsPerTon * fuelTons);        
     }
